Add support for `authn/login-with-expiry` as well as support for `folioRefreshToken` and `folioAccessToken`, and add support for `okapi_login_path`.
Add support for `authn/login-with-expiry` (in theory only, as this is not yet tested).

The login path may be different in future FOLIO versions.
Add support for setting the custom login path via the new configuration setting `okapi_login_path`.
This will fallback to the default `/authn/login` if that configuration setting is not set.

Attempt to determine the authn data based on headers and payload structure in the response.

Folio now uses the terminology "Access Token" in place of the terminology "Okapi Token".
Add support for both the Access Token and the Refresh Token.
Provide new configuration settings called "folioAccessToken" and "folioRefreshToken".
This makes the behavior more consistent with FOLIO terminology.

The original `token` value is preserved for backwards compatibility.