Issue 68: Implement redirect handling.
Perform redirects when receiving HTTP 301 and HTTP 302 status codes.
This implements a hard-coded max redirect of 10.
A custom property could be implemented but this is left for a future decision or for a reviewers input.

This adds new parameters to the RestService HTTP functions for performing the redirets.
The functions are all made to have the same parameters so that the redirect can pass a consistent callback.
Unused parameters, such as `json` in the `get()` method, are ignored.

Logging is added for redirects, when `logLevel` is set to `DEBUG` (case insensitive).
An error is printed when max redirects is reached.

When no redirection is requested, the `redirectRecurse()` function returns false so that the caller can continue normal behavior.