Fix login problem where an error is printed, despite login success, improve success/failure message, and don't print password.
The primary problem is that the token information may be undefined.
The `config.set()` prints the observed error when this variable is undefined.
Add wrapping conditional block to ensure that this error is not triggered on undefined or NULL.

Add additional sanity checks to make sure variable processing does not have undefined or NULL related errors.

It is very unclear on whether or not the login succeeded.
Change the `resolve()` and `reject()` behavior to return a JSON object that has the status message in it.

For security reasons, do not console log the password.
Put the user name in both the login and the failure messages.