Bring in improvements to the RTR support.
This rewrites the RTR token handling logic to better handle cookies.
An entire structure is used now rather than just the token.
The old approach was minimal changes needed to be functional.
This new approach takes a more complete approach to handling.
The `buildAccessHeaders()` constructs the headers for this behavior.

The `fw-cli` is intended to be used in such a way that if the token expires, then the user should just login again.
This avoids needing to implement the refresh support.
Implementing the refresh support would introduce many more changes than is desired and is therefore not implemented.

Bring some of the common error handling functionality to functions to reduce code repetition.
Remove redundant console logging on errors.
There should now be only a single copy of the error being printed.
The error structure being printed is now more consistent.

Example:
```shell
# fw user does_not_exist
{
  status: 'Cannot find user does_not_exist.',
  url: 'https://folio-okapi-dev.library.tamu.edu/users?query=username==does_not_exist',
  payload: undefined,
  http: { code: 200, message: 'OK' },
  body: '{\n' +
    '  "users": [\n' +
    '],\n' +
    '  "totalRecords": 0,\n' +
    '  "resultInfo": {"totalRecords":0,"facets":[],"diagnostics":[]}\n' +
    '}'
}
```

Update the documentation.
There is a lot of missing information that should have been there and now is present.
Remove the registry specific documentation and instead replace it with a link.
This reduces the need to update this project's readme on every registry change.
Remove tamu-specific URLs and data from the example and instead use standard FOLIO example data.

Update the copyrights in all of the files, adding the year 2025.

The mock Okapi server now provides the RTR payloads, with a standard token expirations.
*Note: It is unclear how the mock Okapi server is intended to be run as it appears incomplete.*

Refactor `folioAccessToken` to `accessToken`.
Refactor `folioRefreshToken` to `refreshToken`.
This is done because token is now stored as an object.
For example, the **Access Token** is now at `accessToken.folioAccessToken`.
Storing the entire cookie structure allows for sending the entire cookie structure on requests.

Set defaults to match the FOLIO defaults (such as `admin` to `diku_admin`).
Fix grammar, using proper sentences, in command line help descriptions.

There could be a lot more documentation added to both the readme and the functions.
However, I decided to stop there given the focus and the scope.
The same is true for the console log handling.