Return non-zero status code on error.
This allows for proper scripting and handling.
Non-zero return values generally represents failures.

Use `process.exitCode` in places where there are promises.
This allows for the promise to complete and wrap up even when exiting on error.

Directly call `process.exit()` In other, simpler, cases.

The exit code `1` is being returned for all errors except for the following.
- Promises are instead set to the exit code `2` to represent a promise error.
- Login failure returns exit code `3`.

Improve the error messages, adding `Error:` in front of them and making them complete sentences.
This makes the message more clear and more readable.

The `setup.js` has several awaits with console logs and function calls.
The `process.exitCode` does not trigger an error, so have these cases each only execute if the `process.exitCode` is zero, NULL, or undefined.